+++ webapp/src/components/Diagnosis.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script>
-import Question from '@/modules/Questions/Question.vue'
+import Question from '@/components/Question.vue'
 
   export default {
     name: "Diagnosis",
